Fix POST /business endpoint
Fixed POST /business endpoint. The main issue was with the response from the docClient.post() function, which was expecting the same response as a query() call but instead it actually returns an empty object on success. Tested this end-to-end so once this goes in it should make our listBusiness page work properly